Rename: moved evaluation module into collaborative_filtering module
